library/requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
psutil==5.9.8
pytz==2024.1
sqlparse==0.4.4
